transaction-service/requirements.txt
grpcio 
grpcio-tools
datetime
flask
waitress